test function added

+++ hrc_temperature_plot.perl
@@ -7,36 +7,64 @@ use PGPLOT;
 #												#
 #	author: t. isobe (tisobe@cfa.harvard.edu)						#
 #												#
-#	last update: 09/13/05									#
+#	last update: Feb 05, 2013								#
 #												#
 #################################################################################################
 
 #
+#--- check whether this is a test case
+#
+$comp_test = $ARGV[0];
+chomp $comp_test;
+
+if($comp_test =~ /test/i){
+#
+#--- for the test case
+#
+	$test_out = ' /data/mta/Script/HRC/Temperature/Test_out';
+	system("mkdir $test_out");
+
+	open(OUT, '>./ds_file');
+	print OUT 'columns=mtahrc..hrctemp_avg',"\n";
+	print OUT 'timestart=2010:001:00:00:00',"\n";
+	print OUT 'timestop=2010:060:00:00:00'."\n";
+	close(OUT);
+}else{
+#
 #--- find today's date
 #
 
-($usec, $umin, $uhour, $umday, $umon, $uyear, $uwday, $uyday, $uisdst)= localtime(time);
-$year = $uyear + 1900;
-$month = $umon + 1;
-$today = "$year:$uyday:00:00:00";
+	($usec, $umin, $uhour, $umday, $umon, $uyear, $uwday, $uyday, $uisdst)= localtime(time);
+	$year = $uyear + 1900;
+	$month = $umon + 1;
+	$today = "$year:$uyday:00:00:00";
 
 #
 #--- set dataseeker input file
 #
 
-open(OUT, '>./ds_file');
-print OUT 'columns=mtahrc..hrctemp_avg',"\n";
-print OUT 'timestart=1999:202:00:00:00',"\n";
-print OUT 'timeend='."$today\n";
-close(OUT);
+	open(OUT, '>./ds_file');
+	print OUT 'columns=mtahrc..hrctemp_avg',"\n";
+	print OUT 'timestart=1999:202:00:00:00',"\n";
+	print OUT 'timestop='."$today\n";
+	close(OUT);
+}
 
 #
 #--- call dataseeker
 #
 
 system("punlearn dataseeker; dataseeker.pl infile=ds_file print=yes outfile=hrc_temp.fits");
+
 system("dmlist infile=\"hrc_temp.fits[cols time,2ceahvpt_avg,2chtrpzt_avg,2condmxt_avg,2dcentrt_avg,2dtstatt_avg,2fhtrmzt_avg,2fradpyt_avg,2pmt1t_avg,2pmt2t_avg,2uvlspxt_avg]\" outfile=hrc_temp.dat opt=data");
-system("rm ds_file  hrc_temp.fits");
+
+if($comp_test =~ /test/i){
+	system("mv hrc_temp.fits $test_out");
+	system("gzip $test_out/hrc_temp.fits");
+	system("rm ds_file");
+}else{
+	system("rm ds_file  hrc_temp.fits");
+}
 
 @time = ();
 @ceahvpt = ();
@@ -57,23 +85,11 @@ open(FH, "./hrc_temp.dat");
 OUTER:
 while(<FH>){
 	chomp $_;
-	@atemp = split(/\s+/, $_);
-	if($atemp[12] =~ /\d/){
-		$date = $atemp[2] - 48902399;
-		$date /= 86400;
-		push(@time,    $date);
-		push(@ceahvpt, $atemp[3]);
-		push(@chtrpzt, $atemp[4]);
-		push(@condmxt, $atemp[5]);
-		push(@dcentrt, $atemp[6]);
-		push(@dtstatt, $atemp[7]);
-		push(@fhtrmzt, $atemp[8]);
-		push(@fradpyt, $atemp[9]);
-		push(@pmt1t,   $atemp[10]);
-		push(@pmt2t,   $atemp[11]);
-		push(@uvlspxt, $atemp[12]);
-		$count++;
-	}elsif($atemp[12] eq ''){
+	$string = $_;
+	$string =~ s/^\s+//; 
+	@atemp = split(/\s+/, $string);
+
+	if($atemp[0] =~ /\d/){
 		$date = $atemp[1] - 48902399;
 		$date /= 86400;
 		push(@time,    $date);
@@ -91,11 +107,18 @@ while(<FH>){
 	}
 }
 close(FH);
-system("rm hrc_temp.dat");
+if($comp_test =~ /test/i){
+	system("mv hrc_temp.dat $test_out");
+}else{
+	system("rm hrc_temp.dat");
+}
 
 @temp = sort{$a<=>$b} @time;
-$xmin = $temp[0];
-$xmin = 0;
+if($comp_test =~ /test/){
+	$xmin = $temp[3];
+}else{
+	$xmin = 0;
+}
 $xmax = $temp[$count -1];
 $diff = int($xmax - $xmin);
 $xmax = $xmax + 0.05 * $diff;
@@ -129,6 +152,8 @@ $dtot  = 0;
 #
 #---- modify data interval from 5 min to one day to reduce # of data points
 #
+$start = int($time[0]);
+$end   = $start + 1;
 
 OUTER:
 for($i = 0; $i < $count; $i++){
@@ -146,7 +171,7 @@ for($i = 0; $i < $count; $i++){
 		$scnt++;
 	}elsif($time[$i] < $start){
 		next OUTER;	
-	}elsif($time[$i] > $diff){
+	}elsif($time[$i] > $xmax){
 		last OUTER;
 	}elsif($time[$i] > $end){
 		if($scnt > 0){
@@ -349,10 +374,14 @@ pgptxt($xmid, $ybot, 0.0, 0.0, "Time (DOM)");
 
 pgclos();
 
-$out_plot = '/data/mta_www/mta_hrc/Trending/Temp_data/hrc_temp1.gif';
+if($comp_test =~ /test/i){
+	$out_plot = "$test_out/hrc_temp1.gif";
+}else{
+	$out_plot = '/data/mta_www/mta_hrc/Trending/Temp_data/hrc_temp1.gif';
+}
 
-#system("echo ''|gs -sDEVICE=ppmraw  -r256x256 -q -NOPAUSE -sOutputFile=-  ./pgplot.ps|pnmcrop|pnmflip -r270 |ppmtogif > $out_plot");
-system("echo ''|gs -sDEVICE=ppmraw  -r256x256 -q -NOPAUSE -sOutputFile=-  ./pgplot.ps|pnmflip -r270 |ppmtogif > $out_plot");
+#system("echo ''|/opt/local/bin/gs -sDEVICE=ppmraw  -r256x256 -q -NOPAUSE -sOutputFile=-  ./pgplot.ps|pnmcrop|pnmflip -r270 |ppmtogif > $out_plot");
+system("echo ''|/opt/local/bin/gs -sDEVICE=ppmraw  -r256x256 -q -NOPAUSE -sOutputFile=-  ./pgplot.ps|pnmflip -r270 |ppmtogif > $out_plot");
 system("rm pgplot.ps");
 
 #-----------------------------------------
@@ -502,29 +531,47 @@ pgptxt($xmid, $ybot, 0.0, 0.0, "Time (DOM)");
 pgclos();
 
 
-$out_plot = '/data/mta_www/mta_hrc/Trending/Temp_data/hrc_temp2.gif';
+if($comp_test =~ /test/i){
+	$out_plot = "$test_out/hrc_temp2.gif";
+}else{
+	$out_plot = '/data/mta_www/mta_hrc/Trending/Temp_data/hrc_temp2.gif';
+}
 
-#system("echo ''|gs -sDEVICE=ppmraw  -r256x256 -q -NOPAUSE -sOutputFile=-  ./pgplot.ps|pnmcrop|pnmflip -r270 |ppmtogif > $out_plot");
-system("echo ''|gs -sDEVICE=ppmraw  -r256x256 -q -NOPAUSE -sOutputFile=-  ./pgplot.ps|pnmflip -r270 |ppmtogif > $out_plot");
+#system("echo ''|/opt/local/bin/gs -sDEVICE=ppmraw  -r256x256 -q -NOPAUSE -sOutputFile=-  ./pgplot.ps|pnmcrop|pnmflip -r270 |ppmtogif > $out_plot");
+system("echo ''|/opt/local/bin/gs -sDEVICE=ppmraw  -r256x256 -q -NOPAUSE -sOutputFile=-  ./pgplot.ps|pnmflip -r270 |ppmtogif > $out_plot");
 system("rm pgplot.ps");
 
 open(OUT, '> /data/mta_www/mta_hrc/Trending/hrc_temp.html');
 
-print OUT '<html>',"\n";
-print OUT '<head><title>HRC Temperature</title></head>',"\n";
-print OUT '<body TEXT="#000000" BGCOLOR="#FFFFFF">',"\n";
-print OUT '<center>',"\n";
+
+print OUT "<!DOCTYPE html>\n";
+print OUT "<html>\n";
+print OUT "<head>\n";
+print OUT '<title>HRC Temperature</title>',"\n";
+print OUT "<meta http-equiv='Content-Type' content='text/html; charset=utf-8' />\n";
+print OUT "<style  type='text/css'>\n";
+print OUT "table{text-align:center;margin-left:auto;margin-right:auto;border-style:solid;border-spacing:8px;border-width:2px;border-collapse:separate}\n";
+print OUT "a:link {color:blue;}\n";
+print OUT "a:visited {color:teal;}\n";
+print OUT "</style>\n";
+print OUT "</head>\n";
+
+print OUT "<body style='color:#000000;background-color:#FFFFFF'>\n";
+
+print OUT '<div style="text-align:center;margin-left:auto;margin-right:auto">',"\n";
 print OUT '<h2>Time History of HRC Temperature </h2>',"\n";
-print OUT '</center>',"\n";
+print OUT '</div>',"\n";
 print OUT '',"\n";
-print OUT '<p>',"\n";
+print OUT '<p style="padding-bottom:20px">',"\n";
 print OUT '',"\n";
-print OUT '<img src ="./Temp_data/hrc_temp1.gif" width="800" height="800">',"\n";
-print OUT '<br>',"\n";
-print OUT '<img src ="./Temp_data/hrc_temp2.gif" width="800" height="800">',"\n";
+print OUT '<img src ="./Temp_data/hrc_temp1.gif" alt="hrc temperature 1 plot" style="width:800px;height:800px">',"\n";
+print OUT '<br />',"\n";
+print OUT '<img src ="./Temp_data/hrc_temp2.gif" alt="hrc temperature 2 plot" style="width:800px;height:800px">',"\n";
 print OUT '',"\n";
-print OUT '<br><br>',"\n";
+print OUT '</p>',"\n";
 print OUT '',"\n";
+print OUT "<hr />\n";
+print OUT '<p style="padding-bottom:10px">',"\n";
 
 if($month < 10){
 	$month = '0'."$month";
@@ -534,6 +581,9 @@ if($umday < 10){
 }
 
 print OUT 'Last Update:',"$month/$umday/$year\n";
+print OUT "</p>\n";
+print OUT "</body>\n";
+print OUT "</html>\n";
 
 close(OUT);
 
@@ -548,7 +598,7 @@ $chk = 0;
 while(<FH>){
         chomp $_;
         if($_ =~ /Temperature Time Series Plots/ && $chk == 0){
-                print OUT '<li><a href = "#temperature">Temperature Time Series Plots</a>';
+                print OUT '<li style="font-size:105%;font-weight:bold"><a href = "#temperature">Temperature Time Series Plots</a>';
                 print OUT " (last update: $month-$umday-$year)\n";
                 $chk++;
         }else{
